Correct some errors in french translation
A batiment is masculine in french, so souterrain do not need a 'e', 'feux bricoles' do not exist, that's tricolore. If we have several maison, that's plural. But we have 1 single abri, so no plural form. And vegan is a english word, so that's végane or végan, the former keeping the pronunciation.